+++ locally-app/app/(root)/(tabs)/explore.tsx
