+++ locally-app/app/(root)/_layout.tsx
@@ -1,0 +1,12 @@
+import { Stack } from "expo-router";
+
+const CoreLayout = () => {
+  return (
+    <Stack>
+      <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+      <Stack.Screen name="event-details" options={{ headerShown: false }} />
+    </Stack>
+  );
+};
+
+export default CoreLayout;
