+++ locally-app/app/(root)/event-details.tsx
