+++ locally-app/app/_layout.tsx
@@ -5,11 +5,10 @@ import "../global.css";
 
 export default function RootLayout() {
   return (
-    
     <Stack>
       <Stack.Screen name="index" options={{ headerShown: false }} />
-      <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-      <Stack.Screen name="event-details" options={{ headerShown: false }} />
+      {/* <Stack.Screen name="(auth)" options={{ headerShown: false }} /> */}
+      <Stack.Screen name="(root)" options={{ headerShown: false }} />
     </Stack>
   );
 }
